dev/test/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sre-test"
version = "1.0"
description = "Comprehensive Testing Platform for Spartan Racing Electric"
requires-python = ">=3.8"
dependencies = [
    # Shared
    "click>=8.0.0",
    "rich>=13.0.0",
    "pytest>=7.0.0",
    "pyyaml>=6.0",
    # HIL
    "python-can>=4.0.0",
    "cantools>=36.0.0",
    # SIL
    "pycparser>=2.21",
]

[project.scripts]
sre = "sre_test.cli.main:sre"

[project.optional-dependencies]
dev = [
    "pytest-html>=3.0.0",
    "pytest-cov>=3.0.0",
]